Time: 2 ms (99.84%), Space: 69.5 MB (62.58%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -1,8 +1,8 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
      int left=0;
-     double sum=0;
-     double max=0;
+     int sum=0;
+     int max=0;
      for(int i=0; i<k; i++){
       sum+=nums[i];
      }  
@@ -11,6 +11,6 @@ class Solution {
        sum=sum-nums[left++]+nums[j];
        max=Math.max(sum,max);
      } 
-     return max/k;
+     return (double)max/k;
     }
 }
